refactor, bump version
